+++ main/feature/src/boofcv/alg/feature/detect/interest/FeaturePyramid.java
@@ -202,7 +202,7 @@ public class FeaturePyramid<T extends ImageBase, D extends ImageBase>
 
 			if( checkMax(inten0, val/ss0, x0, y0) && checkMax(inten2, val/ss2, x2, y2) ) {
 				// put features into the scale of the upper image
-				foundPoints.add( new ScalePoint((int)(c.x*scale1),(int)(c.y*scale1),scale1));
+				foundPoints.add( new ScalePoint(c.x*scale1,c.y*scale1,scale1));
 			}
 		}
 	}
